Real email confirmation is here (issue 25). This is both an addition to the Admin CP and Client CP. Admins can now manage unconfirmed emails by going to Clients -> Email Confirmation. Users can check the status of their email or change it in their control panel under the Change Email navigation item. At this point in time having an unconfirmed email doesn't really do anything. In future releases we can use this info to automatically suspend or terminate accounts that have not validated their emails in a certain amount of time. Or suspend accounts immediately after signup until they have confirmed their email, much like Admin Validation is now. Confirm codes are just sha1 hashes of some data. Chance of collision is within acceptable limits. Even if hashes were to collide it wouldn't be disastrous as they're not logically required to be unique. This revision replaces the initial implementation of this idea by kujoez in r53-r55 and r58.
You can check on the status of a user's email from anywhere using the $main->getEmailStatus method. It returns an integer 0 meaning the email is confirmed, 1 meaning "accepted" or not confirmed but email validation is not required, 2 meaning the email is unconfirmed, and 3 meaning the email is unconfirmed because they're attempting to change their email.

When a user's account is initially suspended due to Admin Validation their account is suspended with the reason "TheHostingTool: Awaiting Admin Validation"

Also introduced the ability to re-order the Client CP navigation in the database through a weight column.

git-svn-id: http://thehostingtool.googlecode.com/svn/trunk@286 3066ca77-9623-52e6-b9fb-6191fef3a309